+++ lib/src/header/header.dart
@@ -5,7 +5,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 
 import '../../easy_refresh.dart';
-import '../i18n/global_easy_refresh_localizations.dart';
 
 /// Header
 abstract class Header {
@@ -447,58 +446,33 @@ class ClassicalHeaderWidgetState extends State<ClassicalHeaderWidget>
     }
   }
 
-  /// 默认语言
-  GlobalEasyRefreshLocalizations _localizations =
-      GlobalEasyRefreshLocalizations();
-
   /// 文本
   String get _refreshText {
-    return widget.classicalHeader.refreshText ??
-        (GlobalEasyRefreshLocalizations.of(context) == null
-            ? _localizations.pullToRefresh
-            : GlobalEasyRefreshLocalizations.of(context).pullToRefresh);
+    return widget.classicalHeader.refreshText ?? 'Pull to refresh';
   }
 
   String get _refreshReadyText {
-    return widget.classicalHeader.refreshReadyText ??
-        (GlobalEasyRefreshLocalizations.of(context) == null
-            ? _localizations.releaseToRefresh
-            : GlobalEasyRefreshLocalizations.of(context).releaseToRefresh);
+    return widget.classicalHeader.refreshReadyText ?? 'Release to refresh';
   }
 
   String get _refreshingText {
-    return widget.classicalHeader.refreshingText ??
-        (GlobalEasyRefreshLocalizations.of(context) == null
-            ? _localizations.refreshing
-            : GlobalEasyRefreshLocalizations.of(context).refreshing);
+    return widget.classicalHeader.refreshingText ?? 'Refreshing...';
   }
 
   String get _refreshedText {
-    return widget.classicalHeader.refreshedText ??
-        (GlobalEasyRefreshLocalizations.of(context) == null
-            ? _localizations.refreshed
-            : GlobalEasyRefreshLocalizations.of(context).refreshed);
+    return widget.classicalHeader.refreshedText ?? 'Refresh completed';
   }
 
   String get _refreshFailedText {
-    return widget.classicalHeader.refreshFailedText ??
-        (GlobalEasyRefreshLocalizations.of(context) == null
-            ? _localizations.refreshFailed
-            : GlobalEasyRefreshLocalizations.of(context).refreshFailed);
+    return widget.classicalHeader.refreshFailedText ?? 'Refresh failed';
   }
 
   String get _noMoreText {
-    return widget.classicalHeader.noMoreText ??
-        (GlobalEasyRefreshLocalizations.of(context) == null
-            ? _localizations.noMore
-            : GlobalEasyRefreshLocalizations.of(context).noMore);
+    return widget.classicalHeader.noMoreText ?? 'No more';
   }
 
   String get _infoText {
-    return widget.classicalHeader.infoText ??
-        (GlobalEasyRefreshLocalizations.of(context) == null
-            ? _localizations.updateAt
-            : GlobalEasyRefreshLocalizations.of(context).updateAt);
+    return widget.classicalHeader.infoText ?? 'Update at %T';
   }
 
   // 是否刷新完成