+++ lib/src/refresher.dart
@@ -87,6 +87,7 @@ class EasyRefresh extends StatefulWidget {
 
   /// 全局默认Header
   static Header _defaultHeader = ClassicalHeader();
+
   static set defaultHeader(Header header) {
     if (header != null) {
       _defaultHeader = header;
@@ -95,6 +96,7 @@ class EasyRefresh extends StatefulWidget {
 
   /// 全局默认Footer
   static Footer _defaultFooter = ClassicalFooter();
+
   static set defaultFooter(Footer footer) {
     if (footer != null) {
       _defaultFooter = footer;
@@ -215,6 +217,7 @@ class EasyRefresh extends StatefulWidget {
 class _EasyRefreshState extends State<EasyRefresh> {
   // Physics
   EasyRefreshPhysics _physics;
+
   // Header
   Header get _header {
     if (_enableFirstRefresh && widget.firstRefreshWidget != null)
@@ -224,8 +227,10 @@ class _EasyRefreshState extends State<EasyRefresh> {
 
   // 是否开启首次刷新
   bool _enableFirstRefresh = false;
+
   // 首次刷新组件
   Header _firstRefreshHeader;
+
   // Footer
   Footer get _footer {
     return widget.footer ?? EasyRefresh._defaultFooter;
@@ -243,10 +248,13 @@ class _EasyRefreshState extends State<EasyRefresh> {
 
   // 滚动焦点状态
   ValueNotifier<bool> _focusNotifier;
+
   // 任务状态
   ValueNotifier<TaskState> _taskNotifier;
+
   // 触发刷新状态
   ValueNotifier<bool> _callRefreshNotifier;
+
   // 触发加载状态
   ValueNotifier<bool> _callLoadNotifier;
 
@@ -365,8 +373,10 @@ class _EasyRefreshState extends State<EasyRefresh> {
           growable: true,
         );
       else if (widget.child != null) slivers = _buildSliversByChild();
-      // 判断是否有空视图
-      if (widget.emptyWidget != null && slivers != null) {
+      // 判断是否有空视图(自定义首次刷新视图不用添加，避免视图层叠)
+      if (widget.emptyWidget != null &&
+          slivers != null &&
+          !(_enableFirstRefresh && widget.firstRefreshWidget != null)) {
         slivers = slivers.sublist(0, widget.headerIndex ?? 0);
         // 添加空元素避免异常
         slivers.add(SliverList(
@@ -557,6 +567,7 @@ class EasyRefreshController {
 
   /// 完成刷新
   FinishRefresh finishRefreshCallBack;
+
   void finishRefresh({
     bool success,
     bool noMore,
@@ -568,6 +579,7 @@ class EasyRefreshController {
 
   /// 完成加载
   FinishLoad finishLoadCallBack;
+
   void finishLoad({
     bool success,
     bool noMore,
@@ -579,6 +591,7 @@ class EasyRefreshController {
 
   /// 恢复刷新状态(用于没有更多后)
   VoidCallback resetRefreshStateCallBack;
+
   void resetRefreshState() {
     if (resetRefreshStateCallBack != null) {
       resetRefreshStateCallBack();
@@ -587,6 +600,7 @@ class EasyRefreshController {
 
   /// 恢复加载状态(用于没有更多后)
   VoidCallback resetLoadStateCallBack;
+
   void resetLoadState() {
     if (resetLoadStateCallBack != null) {
       resetLoadStateCallBack();